微修正

+++ src/main/java/org/hiro/input/keyboard/DropCommand.java
@@ -1,11 +1,18 @@
 package org.hiro.input.keyboard;
 
+import org.hiro.Global;
+import org.hiro.IOUtil;
 import org.hiro.ThingMethod;
 
 public class DropCommand implements KeyboardCommand {
 
     @Override
     public void execute() {
-        ThingMethod.drop();
+        if(ThingMethod.dropCheck()){
+            ThingMethod.drop();
+        }else {
+            Global.after = false;
+            IOUtil.msg("there is something there already");
+        }
     }
 }
